vision/requirements.txt
torch>=2.2
torchvision>=0.17
PyYAML>=6.0
Pillow>=10.0
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
